+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Zyp/Moksha/jq: `to_object` function now respects a `zap` option, that
   removes the element altogether if it's empty
 - Zyp/Moksha/jq: Improve error reporting at `MokshaTransformation.apply`
+- MongoDB: Improved `MongoDBCrateDBConverter.decode_extended_json`
 
 ## 2024/09/22 v0.0.17
 - MongoDB: Fixed edge case when decoding MongoDB Extended JSON elements